remove need for device manager, move losses into separate script to reduce clutter
